back/requirements.txt
Django==2.2.17
django-cors-headers==3.0.2
djangorestframework==3.9.4
